Updated with basic get method
