+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
 spring.application.name=SpringbootDemo
 #logging.level.org.springframework=DEBUG
 management.endpoints.web.exposure.include=*
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.jpa.defer-datasource-initialization=true
